优化兼容性

+++ tools/disk.sh
@@ -92,7 +92,7 @@ ask_select(){
         fi
         INPUT=''
         if ((ATTEMPT >= 10));then
-            echo "已经连续输入错误 ${ATTEMPT} 次，终止执行！"
+            echo "已经连续输入错误 ${ATTEMPT} 次，终止询问！"
             return 1
         else
             echo "输入错误，请注意输入选项要求！"
@@ -105,8 +105,6 @@ ask_select(){
         return 1
     fi
 }
-
-
 # 格式化处理
 # @command disk_format $disk
 # @param $part_name             格式化写入分区变量名
@@ -114,50 +112,56 @@ ask_select(){
 # return 1|0
 disk_format(){
     # 获取当前存在的分区
-    local PARTS_PATH=`lsblk -nlp $2|awk '$6 == "part" && !$7{print $1}'` END_SECTOR=$SECTOR_NUM
+    local PARTS_PATH=`lsblk -nl $2|awk '$6 == "part" && !$7{print "/dev/"$1}'` END_SECTOR=$SECTOR_NUM
     echo "即将对磁盘 $2 进行分区，扇区大小：$ARGV_sector_size"
     if [ -n "$END_SECTOR" ];then
-        local START_SECTOR=`fdisk -l /dev/sdb|grep -P '^/dev/'|awk 'END{print $3}'`
-        if [ -n "$START_SECTOR" ];then
-            END_SECTOR=$[$END_SECTOR + $START_SECTOR + 1]
+        if fdisk -l $2|grep -P '^Units'|grep -qP '1\s*\*\s*\d+';then
+            # 按扇区数分区
+            local START_SECTOR=`fdisk -l /dev/sdb|grep -P '^/dev/'|awk 'END{print $3}'`
+            if [ -n "$START_SECTOR" ];then
+                END_SECTOR=$[$END_SECTOR + $START_SECTOR + 1]
+            fi
+            echo "分区大小限制为：$ARGV_part_size ，扇区数：$SECTOR_NUM ，结束扇区数：$END_SECTOR"
+        else
+            # 按柱面数分区
+            END_SECTOR='+'$((PART_SIZE / 1024))'K'
+            echo "分区大小限制为：$ARGV_part_size ，分区指定大小：$END_SECTOR"
         fi
-        echo "分区大小限制为：$ARGV_part_size ，扇区数：$SECTOR_NUM ，结束扇区数：$END_SECTOR"
+    fi
+    local PART_INDEX=`echo -e "1\n2\n3\n4\n$(fdisk -l $2|grep -oP "^$2[1-4]\D"|grep -oP '\d+$')"|sort|grep -P '\d+'|uniq -u|head -n 1`
+    if [ -z "$PART_INDEX" ];then
+        echo "fdisk 分区序号是1~4，当前序号都使用完，无法再分区！"
+        return 1
     fi
     fdisk -b $ARGV_sector_size $2 <<EOF
 n
 p
-
+$PART_INDEX
 
 $END_SECTOR
-w
+wq
 EOF
-    # sleep 2
-    # 格式化完后再获取唯一多出来的分区
-    PARTS_PATH="$PARTS_PATH\n"`lsblk -nlp $2|awk '$6 == "part" && !$7{print $1}'`
-    PART_PATH=$(printf "$PARTS_PATH"|sort|uniq -u)
-    if [ -n "$PART_PATH" ];then
-        echo "成功创建主分区：$PART_PATH，并格式化为：$ARGV_type"
-        mkfs -t $ARGV_type $PART_PATH
-        eval "$1=\$PART_PATH"
-        return 0
-    else
-        echo "分区 $2 失败！"
-        return 1
+    if [ $? = '0' ];then
+        # sleep 2
+        # 格式化完后再获取唯一多出来的分区
+        PARTS_PATH="$PARTS_PATH\n"`lsblk -nl $2|awk '$6 == "part" && !$7{print "/dev/"$1}'`
+        PART_PATH=$(printf "$PARTS_PATH"|sort|uniq -u)
+        if [ -n "$PART_PATH" ];then
+            echo "成功创建主分区：$PART_PATH，并格式化为：$ARGV_type"
+            mkfs -t $ARGV_type $PART_PATH
+            eval "$1=\$PART_PATH"
+            return 0
+        fi
     fi
+    echo "分区 $2 失败！"
+    return 1
 }
-
-
 # 挂载处理
 # @command part_mount $part
 # @param $part                  要挂载的分区地址
 # return 1|0
 part_mount(){
-    PATH_STR="$ARGV_path"
-    if [ -z "$PATH_STR" ];then
-        printf '请输入要挂载的目录：'
-        read PATH_STR
-    fi
-    PART_INFO=(`lsblk -nfl $1|awk '{print $1,$2,$4}'|tail -1`)
+    local PART_INFO=(`lsblk -nfl $1|awk '{print $1,$2,$4}'|tail -1`)
     if ((${#PART_INFO[@]} < 1));then
         echo "无法获取分区信息：$1";
     elif [ -z "${PART_INFO[1]}" ];then
@@ -165,9 +169,25 @@ part_mount(){
     elif [ -n "${PART_INFO[2]}" ];then
         echo "分区 $1 已经挂载到：${PART_INFO[2]}";
     else
-        if [ ! -d "$PATH_STR" ];then
-            mkdir -p $PATH_STR
-        fi
+        local PATH_STR="$ARGV_path" ATTEMPT=1
+        while (true); do
+            if ((ATTEMPT >= 10));then
+                echo "已经连续取消 ${ATTEMPT} 次，终止挂载！"
+                return 1
+            fi
+            if [ -z "$PATH_STR" ];then
+                printf '请输入要挂载的目录：'
+                read PATH_STR
+            fi
+            if [ ! -d "$PATH_STR" ];then
+                mkdir -p $PATH_STR
+            elif [ -n "$(ls -al $PATH_STR|awk '$9!="." && $9!= ".." && $9{print $9}')" ] && ! ask_select "目录 $PATH_STR 不是空的，挂载后会目录内的文件不可访问，确认要挂载？";then
+                PATH_STR=''
+                ((ATTEMPT++))
+                continue
+            fi
+            break
+        done
         if mount $1 $PATH_STR;then
             if [ "$ARGV_temp" != '1' ];then
                 if grep -qP "^$1 " /etc/fstab;then
@@ -192,7 +212,6 @@ part_mount(){
     fi
     return 1
 }
-
 # 容量整理
 # @command size_format $var_name $size
 # @param $var_name              格式化写入变量名
@@ -200,13 +219,13 @@ part_mount(){
 # return 1|0
 size_format(){
     local _SIZE
-    if(($2 > 1099511627776));then
+    if(($2 >= 1099511627776));then
         _SIZE=$[$2 / 1099511627776]'T'
-    elif(($2 > 1073741824));then
+    elif(($2 >= 1073741824));then
         _SIZE=$[$2 / 1073741824]'G'
-    elif(($2 > 1048576));then
+    elif(($2 >= 1048576));then
         _SIZE=$[$2 / 1048576]'M'
-    elif(($2 > 1024));then
+    elif(($2 >= 1024));then
         _SIZE=$[$2 / 1024]'K'
     else
         _SIZE=$2'B'
@@ -265,6 +284,8 @@ Help:
     此脚本只用来快速分区和挂载之用，挂载后默认自动写入 /etc/fstab 配置文件中，以便重启均能生效。
     脚本会自动搜索未挂载的分区或未使用的硬盘，引导去分区和挂载操作。已经挂载的分区将不在此脚本处理范围内。
     此操作会影响硬盘分区，需使用root类最高权限账号操作，操作前请确认操作的必要性。
+
+    注意：脚本强制每个磁盘会留下1M空间给EFI启动代码，所以当磁盘还有1M空间未使用时会自动过滤掉。
 ";
     exit 0
 fi
@@ -299,7 +320,7 @@ if [ "$ARGV_part_size" != '0' ];then
                 SIZE_UNIT=1
                 ;;
         esac
-        PART_SIZE=$[ $SIZE_UNIT * ${ARGV_part_size:0:-1} ]
+        PART_SIZE=$[ $SIZE_UNIT * ${ARGV_part_size/[BKMGT]/} ]
         if (( $ARGV_sector_size > $PART_SIZE ));then
             error_exit "--part-size 分区大小不能小于扇区大小，请核对：$ARGV_part_size"
         else
@@ -318,17 +339,17 @@ DISKS_ARRAY=()
 while read ITEM;do
     BLOCK_ARRAY=($ITEM)
     BLOCK_SIZE=(`lsblk -bna ${BLOCK_ARRAY[0]}|awk 'BEGIN{use=0;part=0}$6 != "disk" && $7 != ""{use += $4} $6 == "part"{part += $4}END{print use,part}'`)
-    if (( ${BLOCK_ARRAY[3]} <= ${BLOCK_SIZE[0]} ));then
+    if (( ${BLOCK_ARRAY[3]} <= ${BLOCK_SIZE[0]} )) || ([ "${BLOCK_ARRAY[5]}" = 'part' ] && (( ${BLOCK_SIZE[0]} > 0 )));then
         continue
     elif [ "${BLOCK_ARRAY[5]}" = 'disk' ];then
         BLOCK_ARRAY[3]=$[ ${BLOCK_ARRAY[3]} - ${BLOCK_SIZE[1]} ]
-        if (( ${BLOCK_ARRAY[3]} <= 0 ));then
+        if (( ${BLOCK_ARRAY[3]}/1048576 <= 1 ));then
             continue
         fi
     fi
     DISKS_ARRAY[${#DISKS_ARRAY[@]}]="${BLOCK_ARRAY[@]}"
 done <<EOF
-`lsblk -bnapl $ARGV_disk|awk '$6 ~ "disk|part" && !$7 ""{print}'`
+`lsblk -bnal $ARGV_disk|awk '$6 ~ "disk|part" && !$7 {$1=("/dev/"$1);print}'`
 EOF
 
 if ((${#DISKS_ARRAY[@]} < 1));then
@@ -347,7 +368,7 @@ for ((INDEX=0;INDEX<${#DISKS_ARRAY[@]};INDEX++));do
     fi
     if [ ${ARRAY_DATA[5]} = 'disk' ];then
         if ask_select "硬盘 ${ARRAY_DATA[0]} 未用空间：$UNUSERD_SIZE 是否格式化再挂载？" && disk_format PART_NAME ${ARRAY_DATA[0]};then
-            PART_INFO=(`lsblk -nlp $PART_NAME|awk '$6 == "part" && $7 == "" {print $1,$4}'|tail -1`)
+            PART_INFO=(`lsblk -nl $PART_NAME|awk '$6 == "part" && $7 == "" {print "/dev/"$1,$4}'|tail -1`)
             echo "硬盘 ${ARRAY_DATA[0]} 刚创建分区 ${PART_INFO[0]} 可用空间：${PART_INFO[1]} 即将挂载"
             if part_mount ${PART_INFO[0]} && [ -n "$ARGV_path" ];then
                 exit 0
